+++ Plugins/GeoModel/src/detail/GeoIntersectionAnnulusConverter.cpp
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "Acts/Plugins/GeoModel/converters/GeoIntersectionAnnulusConverter.hpp"
+#include "Acts/Plugins/GeoModel/detail/GeoIntersectionAnnulusConverter.hpp"
 
 #include "Acts/Definitions/Common.hpp"
 #include "Acts/Definitions/Units.hpp"
