+++ Plugins/GeoModel/src/detail/GeoShiftConverter.cpp
@@ -6,7 +6,15 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "Acts/Plugins/GeoModel/converters/GeoShiftConverter.hpp"
+#include "Acts/Plugins/GeoModel/detail/GeoShiftConverter.hpp"
+
+#include "Acts/Plugins/GeoModel/GeoModelConversionError.hpp"
+#include "Acts/Plugins/GeoModel/detail/GeoBoxConverter.hpp"
+#include "Acts/Plugins/GeoModel/detail/GeoTrdConverter.hpp"
+#include "Acts/Plugins/GeoModel/detail/GeoTubeConverter.hpp"
+#include "Acts/Surfaces/PlaneSurface.hpp"
+#include "Acts/Surfaces/StrawSurface.hpp"
+#include "Acts/Surfaces/TrapezoidBounds.hpp"
 
 namespace Acts::detail {
