+++ Plugins/GeoModel/src/detail/GeoTrdConverter.cpp
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "Acts/Plugins/GeoModel/converters/GeoTrdConverter.hpp"
+#include "Acts/Plugins/GeoModel/detail/GeoTrdConverter.hpp"
 
 #include "Acts/Definitions/Common.hpp"
 #include "Acts/Definitions/Units.hpp"
